style: 함수, useState, useEffect 부모로 올리기
